+++ packages/eyes-storybook/src/getStoryUrl.js
@@ -2,15 +2,19 @@
 const getIframeUrl = require('./getIframeUrl');
 
 function getStoryUrl({name, kind, parameters}, baseUrl) {
-  const queryParams = (parameters && parameters.eyes && parameters.eyes.queryParams) || {};
-  const queryString = Object.entries(queryParams).reduce(
-    (queryString, [name, value]) => queryString + `&${name}=${value}`,
-    '',
-  );
+  let storyUrl = getIframeUrl(baseUrl);
+  storyUrl += `&selectedKind=${encodeURIComponent(kind)}`;
+  storyUrl += `&selectedStory=${encodeURIComponent(name)}`;
 
-  return `${getIframeUrl(baseUrl)}&selectedKind=${encodeURIComponent(
-    kind,
-  )}&selectedStory=${encodeURIComponent(name)}${queryString}`;
+  const queryParams = parameters && parameters.eyes && parameters.eyes.queryParams;
+  if (queryParams) {
+    storyUrl += `&eyes-query-params=${Object.keys(queryParams).join(',')}`;
+    for (const [name, value] of Object.entries(queryParams)) {
+      storyUrl += `&${name}=${encodeURIComponent(value)}`;
+    }
+  }
+
+  return storyUrl;
 }
 
 module.exports = getStoryUrl;